django/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-rest-auth==2.1.3
Django==3.1.5
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.13
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
nodeenv==1.5.0
packaging==20.9
pre-commit==2.10.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
